Build on msvc x86-64.

+++ components/media/adobe/xfl/tests/xfl_load.cpp
@@ -39,7 +39,7 @@ void print (const math::vector<T, N>& v)
 {
   printf ("[");
 
-  for (size_t i=0; i<N; i++)
+  for (unsigned int i=0; i<N; i++)
   {
     printf (" ");
     print  (v [i]);
